Extract non-value-setting logic from `TyneFilterValueBase` into `TyneFilterValueCore`
